*** empty log message ***
git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@10530 678b6752-a8a0-49ff-a511-7c2a3241fcd4

+++ betarun/CRUN/ExtendRefRep.c
@@ -37,7 +37,7 @@ ParamObjOffRange(ExtRR)
     size = RefRepSize(newRange);
     do {
       if (newRange>LARGE_REP_SIZE || size>IOAMAXSIZE){
-	DEBUG_AOA(fprintf(output, "ExtRR allocates in AOA(newRange=%d)\n", newRange));
+	DEBUG_AOA(fprintf(output, "ExtRR allocates in AOA(newRange=%d)\n", (int)newRange));
 	newRep = (RefRep *)AOAcalloc(size);
 	DEBUG_AOA(if (!newRep) fprintf(output, "AOAcalloc failed\n"));
       }
